Adiciona uma abordagem utilizando o pattern Strategy

+++ src/main/java/br/com/alelo/consumer/consumerpat/entity/Extract.java
@@ -17,11 +17,11 @@ public class Extract {
     String establishmentName;
     String productDescription;
     Date dateBuy;
-    int cardNumber;
+    Long cardNumber;
     double value;
 
 
-    public Extract(int id, int establishmentNameId, String establishmentName, String productDescription, Date dateBuy, int cardNumber, double value) {
+    public Extract(int id, int establishmentNameId, String establishmentName, String productDescription, Date dateBuy, Long cardNumber, double value) {
         this.id = id;
         this.establishmentNameId = establishmentNameId;
         this.establishmentName = establishmentName;
@@ -31,14 +31,14 @@ public class Extract {
         this.value = value;
     }
 
-    public Extract( String productDescription, Date dateBuy, int cardNumber, double value) {
+    public Extract( String productDescription, Date dateBuy, Long cardNumber, double value) {
         this.productDescription = productDescription;
         this.dateBuy = dateBuy;
         this.cardNumber = cardNumber;
         this.value = value;
     }
 
-    public Extract(String establishmentName, String productDescription, Date dateBuy, int cardNumber, double value) {
+    public Extract(String establishmentName, String productDescription, Date dateBuy, Long cardNumber, double value) {
         this.establishmentNameId = establishmentNameId;
         this.establishmentName = establishmentName;
         this.productDescription = productDescription;
